Move views to sub-package
