+++ src/main/java/com/vlf/intentsender/ui/views/ExtrasDeleteButtonEditor.java
@@ -1,4 +1,4 @@
-package com.vlf.intentsender.ui;
+package com.vlf.intentsender.ui.views;
 
 import java.awt.Component;
 import java.awt.event.ActionEvent;
@@ -12,7 +12,7 @@ import javax.swing.JTable;
 /**
  * Created by vfarafonov on 31.08.2015.
  */
-class ExtrasDeleteButtonEditor extends DefaultCellEditor {
+public class ExtrasDeleteButtonEditor extends DefaultCellEditor {
 
 
 	private final JButton button_;
