+++ src/main/java/com/vlf/intentsender/ui/views/HistoryListCellRenderer.java
@@ -1,4 +1,4 @@
-package com.vlf.intentsender.ui;
+package com.vlf.intentsender.ui.views;
 
 import java.awt.Component;
 
@@ -10,7 +10,7 @@ import com.vlf.intentsender.Models.Command;
 /**
  * Created by vfarafonov on 08.09.2015.
  */
-class HistoryListCellRenderer extends DefaultListCellRenderer {
+public class HistoryListCellRenderer extends DefaultListCellRenderer {
 	private static final int MAX_SUBJECT_LENGTH = 30;
 
 	@Override
